executor_binary.c has up to date working pipeline
