+++ src/utils.c
@@ -6,7 +6,7 @@
 /*   By: tkajanek <tkajanek@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/03 18:05:14 by tkajanek          #+#    #+#             */
-/*   Updated: 2023/06/15 16:14:52 by tkajanek         ###   ########.fr       */
+/*   Updated: 2023/06/17 18:21:46 by tkajanek         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,27 @@ int	commands_counting(char **words)
     return (count);
 }
 
-void free_args(char **args)
+void	free_args(char ***args)
+{
+	int i;
+	int j;
+	
+	i = 0;
+    while (args[i] != NULL)
+	{
+        j = 0;
+        while (args[i][j] != NULL)
+		{
+            free(args[i][j]);
+            j++;
+        }
+        free(args[i]);
+        i++;
+    }
+    free(args);
+}
+
+void	free_split(char **args)
 {
     int i;
